pyproject.toml
[project]
name = "query-to-pdf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.15"
dependencies = [
    "aiohttp>=3.8.0",
    "asyncio>=3.4.3",
    "asyncio-throttle>=1.0.2",
    "beautifulsoup4>=4.12.0",
    "firecrawl-py>=2.9.0",
    "google-api-python-client>=2.100.0",
    "google-auth>=2.23.0",
    "google-auth-httplib2>=0.1.1",
    "google-auth-oauthlib>=1.1.0",
    "google-generativeai>=0.5.0",
    "jinja2>=3.1.0",
    "kaleido>=0.2.1",
    "lxml>=4.9.0",
    "matplotlib>=3.7.0",
    "numpy>=1.24.0",
    "openai>=1.3.0",
    "pandas>=2.0.0",
    "pillow>=10.0.0",
    "plotly>=5.17.0",
    "pydantic>=2.0.0",
    "pytest>=7.4.0",
    "python-dotenv>=1.0.0",
    "reportlab>=4.0.4",
    "requests>=2.31.0",
    "scipy>=1.11.0",
    "seaborn>=0.12.0",
    "streamlit>=1.28.0",
    "weasyprint>=60.0",
]
